refactor: make Grid a compound component
