+++ packages/core/src/Grid/index.ts
@@ -1,4 +1,4 @@
-export { Grid } from './Grid';
+export { Grid } from './CompoundGrid';
 export type { GridProps } from './Grid';
 export { GridItem } from './GridItem/GridItem';
 export type { GridItemProps } from './GridItem/GridItem';
